Update Cal and add UI customizations

+++ pages/api/availability/eventtype.ts
@@ -1,9 +1,51 @@
+import { EventTypeCustomInput, MembershipRole, Prisma } from "@prisma/client";
 import type { NextApiRequest, NextApiResponse } from "next";
+
 import { getSession } from "@lib/auth";
 import prisma from "@lib/prisma";
+import { OpeningHours } from "@lib/types/event-type";
+
+function handleCustomInputs(customInputs: EventTypeCustomInput[], eventTypeId: number) {
+  if (!customInputs || !customInputs?.length) return undefined;
+  const cInputsIdsToDelete = customInputs.filter((input) => input.id > 0).map((e) => e.id);
+  const cInputsToCreate = customInputs
+    .filter((input) => input.id < 0)
+    .map((input) => ({
+      type: input.type,
+      label: input.label,
+      required: input.required,
+      placeholder: input.placeholder,
+    }));
+  const cInputsToUpdate = customInputs
+    .filter((input) => input.id > 0)
+    .map((input) => ({
+      data: {
+        type: input.type,
+        label: input.label,
+        required: input.required,
+        placeholder: input.placeholder,
+      },
+      where: {
+        id: input.id,
+      },
+    }));
+
+  return {
+    deleteMany: {
+      eventTypeId,
+      NOT: {
+        id: { in: cInputsIdsToDelete },
+      },
+    },
+    createMany: {
+      data: cInputsToCreate,
+    },
+    update: cInputsToUpdate,
+  };
+}
 
 export default async function handler(req: NextApiRequest, res: NextApiResponse) {
-  const session = await getSession({ req: req });
+  const session = await getSession({ req });
 
   if (!session) {
     res.status(401).json({ message: "Not authenticated" });
@@ -20,6 +62,16 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
       where: { id: req.body.id },
       include: {
         users: true,
+        team: {
+          select: {
+            members: {
+              select: {
+                userId: true,
+                role: true,
+              },
+            },
+          },
+        },
       },
     });
 
@@ -27,67 +79,49 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
       return res.status(404).json({ message: "No event exists matching that id." });
     }
 
-    const isAuthorized =
-      event.userId === session.user.id ||
-      event.users.find((user) => {
-        return user.id === session.user?.id;
-      });
+    const isAuthorized = (function () {
+      if (event.team) {
+        return event.team.members
+          .filter((member) => member.role === MembershipRole.OWNER)
+          .map((member) => member.userId)
+          .includes(session.user.id);
+      }
+      return (
+        event.userId === session.user.id ||
+        event.users.find((user) => {
+          return user.id === session.user?.id;
+        })
+      );
+    })();
 
     if (!isAuthorized) {
       console.warn(`User ${session.user.id} attempted to an access an event ${event.id} they do not own.`);
-      return res.status(404).json({ message: "No event exists matching that id." });
+      return res.status(403).json({ message: "No event exists matching that id." });
     }
   }
 
   if (req.method == "PATCH" || req.method == "POST") {
-    const data = {
+    const data: Prisma.EventTypeCreateInput | Prisma.EventTypeUpdateInput = {
       title: req.body.title,
       slug: req.body.slug.trim(),
       description: req.body.description,
       length: parseInt(req.body.length),
       hidden: req.body.hidden,
       requiresConfirmation: req.body.requiresConfirmation,
+      disableGuests: req.body.disableGuests,
       locations: req.body.locations,
       eventName: req.body.eventName,
-      customInputs: !req.body.customInputs
-        ? undefined
-        : {
-            deleteMany: {
-              eventTypeId: req.body.id,
-              NOT: {
-                id: { in: req.body.customInputs.filter((input) => !!input.id).map((e) => e.id) },
-              },
-            },
-            createMany: {
-              data: req.body.customInputs
-                .filter((input) => !input.id)
-                .map((input) => ({
-                  type: input.type,
-                  label: input.label,
-                  required: input.required,
-                  placeholder: input.placeholder,
-                })),
-            },
-            update: req.body.customInputs
-              .filter((input) => !!input.id)
-              .map((input) => ({
-                data: {
-                  type: input.type,
-                  label: input.label,
-                  required: input.required,
-                  placeholder: input.placeholder,
-                },
-                where: {
-                  id: input.id,
-                },
-              })),
-          },
+      customInputs: handleCustomInputs(req.body.customInputs as EventTypeCustomInput[], req.body.id),
       periodType: req.body.periodType,
       periodDays: req.body.periodDays,
       periodStartDate: req.body.periodStartDate,
       periodEndDate: req.body.periodEndDate,
       periodCountCalendarDays: req.body.periodCountCalendarDays,
-      minimumBookingNotice: req.body.minimumBookingNotice,
+      minimumBookingNotice: req.body.minimumBookingNotice
+        ? parseInt(req.body.minimumBookingNotice)
+        : undefined,
+      price: req.body.price,
+      currency: req.body.currency,
     };
 
     if (req.body.schedulingType) {
@@ -105,10 +139,10 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
 
       const eventType = await prisma.eventType.create({
         data: {
-          ...data,
+          ...(data as Prisma.EventTypeCreateInput),
           users: {
             connect: {
-              id: parseInt(session.user.id),
+              id: session?.user?.id,
             },
           },
         },
@@ -118,7 +152,7 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
       if (req.body.users) {
         data.users = {
           set: [],
-          connect: req.body.users.map((id: number) => ({ id })),
+          connect: req.body.users.map((id: string) => ({ id: parseInt(id) })),
         };
       }
 
@@ -127,28 +161,29 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
       }
 
       if (req.body.availability) {
-        const openingHours = req.body.availability.openingHours || [];
+        const openingHours: OpeningHours[] = req.body.availability.openingHours || [];
         // const overrides = req.body.availability.dateOverrides || [];
 
-        await prisma.availability.deleteMany({
-          where: {
-            eventTypeId: +req.body.id,
+        const eventTypeId = +req.body.id;
+        if (eventTypeId) {
+          await prisma.availability.deleteMany({
+            where: {
+              eventTypeId,
+            },
+          });
+        }
+
+        const availabilityToCreate = openingHours.map((openingHour) => ({
+          startTime: openingHour.startTime,
+          endTime: openingHour.endTime,
+          days: openingHour.days,
+        }));
+
+        data.availability = {
+          createMany: {
+            data: availabilityToCreate,
           },
-        });
-        Promise.all(
-          openingHours.map((schedule) =>
-            prisma.availability.create({
-              data: {
-                eventTypeId: +req.body.id,
-                days: schedule.days,
-                startTime: schedule.startTime,
-                endTime: schedule.endTime,
-              },
-            })
-          )
-        ).catch((error) => {
-          console.log(error);
-        });
+        };
       }
 
       const eventType = await prisma.eventType.update({
